Assignment 7 Nonlinear pendulum & Lotka Volterra
In part b, the populations of the rabbits and foxes are cycling, and as
the foxes population rises, the rabbit population falls. Once the rabbit
population gets low enough, the foxes start to die out as they do not
have as much food as before, and then the fox population delcines, and
as a result the rabbit population increases, then the cycle repeats
itself.